Refactored all imports of urlGenerators.
