+++ web-app/src/components/header/employer-nav-links.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { urlGenerators } from "src/pages/url-generators";
+import * as urlGenerators from "src/pages/url-generators";
 import { translations } from "src/translations";
 
 const T = translations.header;
