+++ web-app/src/pages/create-company/create-company-form.tsx
@@ -8,7 +8,7 @@ import ReactGA from "react-ga";
 
 import { useCreateCompanyMutation } from "generated/graphql-operations";
 import * as schemas from "src/form-schemas";
-import { urlGenerators } from "src/pages/url-generators";
+import * as urlGenerators from "src/pages/url-generators";
 
 import InnerForm from "./create-company-inner-form";
 
@@ -97,7 +97,9 @@ const onSubmit = (createCompany, history, setSubmissionError) => (
 
 			// Go to the newly created company's page.
 			history.push(
-				urlGenerators.vizeCompanyProfileUrl(data.createCompany.company.id)
+				urlGenerators.vizeCompanyProfileUrl(
+					data.createCompany.company.id
+				)
 			);
 		})
 		.catch(errors => {
