+++ web-app/src/pages/create-job-ad/create-job-ad-form.tsx
@@ -5,7 +5,7 @@ import * as yup from "yup";
 import { mapValues, map, omitBy, filter } from "lodash";
 
 import { useCreateJobAdMutation } from "generated/graphql-operations";
-import { urlGenerators } from "src/pages/url-generators";
+import * as urlGenerators from "src/pages/url-generators";
 import ReactPixel from "react-facebook-pixel";
 import ReactGA from "react-ga";
