+++ web-app/src/pages/index.ts
@@ -1,2 +1,1 @@
 export { default } from "./pages";
-export { urlGenerators } from "./url-generators";
